+++ changelog.md
@@ -8,3 +8,7 @@
 - Prompt: Automatic pre-commit tracked update
 - Automatic pre-commit tracking of changed files
 
+## v3 - 2026-05-10 18:59:04
+- Prompt: Automatic pre-commit tracked update
+- Automatic pre-commit tracking of changed files
+